closes #3, fixes #4
Inactive PM is now stored.  Undefined indexes for all unset settings are
now fixed.  Default settings have been created for PM and Email subject
and message.  All other settings are checked and either don't allow the
list to be created and give a message to the user setup the missing
setting or run with a default that allows the list to be run without
errors.